改造

+++ src/main/java/com/wage/controller/DepartmentController.java
@@ -1,11 +1,16 @@
 package com.wage.controller;
 
 import com.github.pagehelper.PageHelper;
+import com.wage.model.Admin;
 import com.wage.model.Department;
+import com.wage.model.Employee;
+import com.wage.service.AdminService;
 import com.wage.service.DepartmentService;
 import com.github.pagehelper.PageInfo;
+import com.wage.service.EmployeeService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Resource;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -27,6 +33,12 @@ public class DepartmentController {
     @Resource
     private DepartmentService departmentService;
 
+    @Resource
+    private EmployeeService employeeService;
+
+    @Resource
+    private AdminService adminService;
+
     /**
      * @Description: 分页查询
      * @param page 页码
@@ -38,6 +50,7 @@ public class DepartmentController {
                        @RequestParam(defaultValue = "0") Integer size, Model model) throws Exception {
         PageHelper.startPage(page, size);
         List<Department> list = departmentService.selectAll();
+        list = list == null ? new ArrayList<>() : list;
         PageInfo<Department> pageInfo = new PageInfo<Department>(list);
         model.addAttribute("pageInfo", pageInfo);
         return "departmentList";
@@ -50,8 +63,8 @@ public class DepartmentController {
 
     @PostMapping("/add")
     public String add(Department department, Model model) throws Exception {
-        Integer state = departmentService.insert(department);
-        if (state == 0){
+        department = departmentService.insert(department);
+        if (department == null){
             model.addAttribute("message", "添加部门失败");
             return "message";
         }else {
@@ -61,7 +74,7 @@ public class DepartmentController {
     }
 
     @GetMapping("/updatePage")
-    public String updatePage(Model model, String id) throws Exception {
+    public String updatePage(Model model, Integer id) throws Exception {
         Department department = departmentService.selectById(id);
         model.addAttribute("department", department);
         return "departmentUpdate";
@@ -69,8 +82,8 @@ public class DepartmentController {
 
     @PostMapping("/update")
     public String update(Department department, Model model) throws Exception {
-        Integer state = departmentService.update(department);
-        if (state == 0){
+        department = departmentService.update(department);
+        if (department == null){
             model.addAttribute("message", "修改部门失败");
             return "message";
         }else {
@@ -80,53 +93,19 @@ public class DepartmentController {
     }
 
     @GetMapping("/delete")
-    public String delete(String id, Model model) throws Exception {
-        Integer state = departmentService.deleteById(id);
-        if (state == 0){
-            model.addAttribute("message", "删除部门失败");
-            return "message";
-        }else {
+    public String delete(Integer id, Model model) throws Exception {
+
+        List<Employee> list = employeeService.findEmployeeByDepartmentId(id);
+        List<Admin> alist = adminService.findAdminsByDepartmentId(id);
+        if(!CollectionUtils.isEmpty(list)) {
+            model.addAttribute("message", "删除部门失败，有员工绑定！");
+        } else if(!CollectionUtils.isEmpty(alist)){
+            model.addAttribute("message", "删除部门失败，有管理员绑定！");
+        } else {
+            departmentService.deleteById(id);
             model.addAttribute("message", "删除部门成功");
-            return "message";
         }
+        return "message";
     }
 
-    /*@PostMapping("/insert")
-    public Result<Integer> insert(Department department) throws Exception{
-        Integer state = departmentService.insert(department);
-        return ResultUtil.SUCCESS(state);
-    }
-
-    @PostMapping("/deleteById")
-    public Result<Integer> deleteById(@RequestParam String id) throws Exception {
-        Integer state = departmentService.deleteById(id);
-        return ResultUtil.SUCCESS(state);
-    }
-
-    @PostMapping("/update")
-    public Result<Integer> update(Department department) throws Exception {
-        Integer state = departmentService.update(department);
-        return ResultUtil.SUCCESS(state);
-    }
-
-    @GetMapping("/selectById")
-    public Result<Department> selectById(@RequestParam String id) throws Exception {
-        Department department = departmentService.selectById(id);
-        return ResultUtil.SUCCESS(department);
-    }
-
-    *//*
-    * @Description: 分页查询
-    * @param page 页码
-    * @param size 每页条数
-    * @Reutrn RetResult<PageInfo<Department>>
-    *//*
-    @GetMapping("/list")
-    public Result<PageInfo<Department>> list(@RequestParam(defaultValue = "0") Integer page,
-            @RequestParam(defaultValue = "0") Integer size) throws Exception {
-        PageHelper.startPage(page, size);
-        List<Department> list = departmentService.selectAll();
-        PageInfo<Department> pageInfo = new PageInfo<Department>(list);
-        return ResultUtil.SUCCESS(pageInfo);
-    }*/
 }
